Convert box-label-identifier unit test into endpoint test
The underlying save_creation_to_history() function requires a running
Flask app context. Instead of mocking/patching it in the unit test I
decided to move the entire test to a higher-level endpoint test. The
covered functionality remains unchanged.
Another test is removed but replaced by an endpoint test in the process
of the on-going test-plan update.
